+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   - Re-style configuration not changed report to look like other reports.
   - Do not show warning when config not changes
   - Add backup directory option
+* [dropwizard]
+  - Add --backup-dir option 
 
 ### 1.2.0 (2021-10-09)
 * [api]